refactor : elasticsearch 미사용 조회 삭제
