+++ docker-compose.yml
@@ -6,10 +6,11 @@ services:
   redis:
     image: redis:6
     container_name: redis
+    environment:
+      REDIS_PASSWORD: $REDIS_PASSWORD
+    command: [ "redis-server", "--requirepass", "${REDIS_PASSWORD}" ]
     ports:
-      - "6379:6379"
-    networks:
-      - spring-net
+     - "6379:6379"
 
   Kafka00Service:
     image: bitnami/kafka:3.7.0
